+++ src/main/java/com/karakurt/homework3/DataAccess/Product/ProductDao.java
@@ -25,48 +25,48 @@ public class ProductDao implements IProductDao {
     public List<IProduct> getRandomProducts() {
         Random random = new Random();
         List<IProduct> randomProducts = new ArrayList<>();
-        randomProducts.add(new Product(1,"Jeans",random.nextDouble(90) + 20));
-        randomProducts.add(new Product(2,"Jacket",random.nextDouble(90) + 20));
-        randomProducts.add(new Product(3,"Poncho",random.nextDouble(90) + 20));
-        randomProducts.add(new Product(4,"Cravat",random.nextDouble(40) + 5));
-        randomProducts.add(new Product(5,"Belt",random.nextDouble(65) + 10));
-        randomProducts.add(new Product(6,"Cheesecake",random.nextDouble(70) + 40));
-        randomProducts.add(new Product(7,"Burritos",random.nextDouble(90) + 50));
-        randomProducts.add(new Product(8,"Sandwich",random.nextDouble(30) + 20));
-        randomProducts.add(new Product(9,"Calzone",random.nextDouble(110) + 80));
-        randomProducts.add(new Product(10,"Chicken",random.nextDouble(70) + 40));
-        randomProducts.add(new Product(11,"Asus",random.nextDouble(400) + 200));
-        randomProducts.add(new Product(12,"Lenovo",random.nextDouble(400) + 200));
-        randomProducts.add(new Product(13,"Iphone",random.nextDouble(400) + 200));
-        randomProducts.add(new Product(14,"Samsung",random.nextDouble(400) + 200));
-        randomProducts.add(new Product(15,"Casper",random.nextDouble(400) + 200));
+        randomProducts.add(new Product(1L,"Jeans",random.nextDouble(90) + 20));
+        randomProducts.add(new Product(2L,"Jacket",random.nextDouble(90) + 20));
+        randomProducts.add(new Product(3L,"Poncho",random.nextDouble(90) + 20));
+        randomProducts.add(new Product(4L,"Cravat",random.nextDouble(40) + 5));
+        randomProducts.add(new Product(5L,"Belt",random.nextDouble(65) + 10));
+        randomProducts.add(new Product(6L,"Cheesecake",random.nextDouble(70) + 40));
+        randomProducts.add(new Product(7L,"Burritos",random.nextDouble(90) + 50));
+        randomProducts.add(new Product(8L,"Sandwich",random.nextDouble(30) + 20));
+        randomProducts.add(new Product(9L,"Calzone",random.nextDouble(110) + 80));
+        randomProducts.add(new Product(10L,"Chicken",random.nextDouble(70) + 40));
+        randomProducts.add(new Product(11L,"Asus",random.nextDouble(400) + 200));
+        randomProducts.add(new Product(12L,"Lenovo",random.nextDouble(400) + 200));
+        randomProducts.add(new Product(13L,"Iphone",random.nextDouble(400) + 200));
+        randomProducts.add(new Product(14L,"Samsung",random.nextDouble(400) + 200));
+        randomProducts.add(new Product(15L,"Casper",random.nextDouble(400) + 200));
         return randomProducts;
     }
 
     @Override
     public IProduct createProduct(String name, Double price) {
         controlRandomProducts();
-        Integer id = randomProducts.size() + 1;
+        Long id = Long.valueOf(randomProducts.size() + 1);
         IProduct iProduct = new Product(id, name, price);
         randomProducts.add(iProduct);
         return iProduct;
     }
 
     @Override
-    public IProduct updateProduct(Integer id, String name, Double price) {
+    public IProduct updateProduct(Long id, String name, Double price) {
         controlRandomProducts();
-        IProduct iProduct = randomProducts.get(id - 1);
+        IProduct iProduct = randomProducts.get(Math.toIntExact(id - 1));
         iProduct.setName(name);
         iProduct.setPrice(price);
-        randomProducts.remove(id - 1);
-        randomProducts.add(id - 1, iProduct);
+        randomProducts.remove(Math.toIntExact(id - 1));
+        randomProducts.add(Math.toIntExact(id - 1), iProduct);
         return iProduct;
     }
 
     @Override
-    public void deleteProduct(Integer id) {
+    public void deleteProduct(Long id) {
         controlRandomProducts();
-        randomProducts.remove(id - 1);
+        randomProducts.remove(Math.toIntExact(id - 1));
     }
 
     public void setRandomProducts(List<IProduct> randomProductList) {
