+++ src/main/java/com/karakurt/homework3/Entities/Product/IProduct.java
@@ -1,8 +1,8 @@
 package com.karakurt.homework3.Entities.Product;
 
 public interface IProduct {
-    Integer getId();
-    void setId(Integer id);
+    Long getId();
+    void setId(Long id);
     String getName();
     void setName(String name);
     Double getPrice();
